starter-apps/fastapi-apps/Multilingual-Budget-Analyzer/fastapi_backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.1.0
plotly>=5.17.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
langchain>=0.1.12
langchain-openai>=0.1.1
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
xlrd>=2.0.0
tabulate>=0.9.0 